refactor: split hover
