+++ packages/global/src/utils/hover/types.ts
@@ -1,0 +1,24 @@
+import type { SafeTriangleOptions } from './safe-area.js';
+
+export type WhenHoverOptions = {
+  enterDelay?: number;
+  leaveDelay?: number;
+  /**
+   * When already hovered to the reference element,
+   * but the floating element is not ready,
+   * the callback will still be executed if the `alwayRunWhenNoFloating` is true.
+   *
+   * It is useful when the floating element is removed just before by a user's action,
+   * and the user's mouse is still hovering over the reference element.
+   *
+   * @default true
+   */
+  alwayRunWhenNoFloating?: boolean;
+  safeArea?: boolean | SafeTriangleOptions;
+};
+
+export type HoverMiddleware = (ctx: {
+  event: Event;
+  referenceElement?: Element;
+  floatingElement?: Element;
+}) => boolean | Promise<boolean>;
